backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.1
playwright>=1.49.0
apscheduler==3.10.4
python-multipart==0.0.12
pydantic==2.9.2
aiofiles==24.1.0
httpx==0.27.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
